#564 and #620 a.	For IBA and IBE, add the alternate labels ‘Information Carrying Artifact’ and ‘Information Carrying Entity’, respectively, and add the comment to IBE: ‘This class continues to use the work ‘bearing’ in its rdfs:label because of potential confusion with the acronym ‘ICE’, which is widely used for information content entity.’
b.	Add a similar comment for IBA that indicates by “bears” we mean that the IBA ‘is carrier of’ an ICE.

c.	IBA should get the comment: “This class continues to use the word ‘bearing’ in its rdfs:label for legacy reasons. It is understood that an IBA bears the concretization of an ICE and not the ICE directly.”
d.	Update all rdfs:labels of subclasses of IBA to use prefix ‘Material Copy of’ and make their existing rdfs:label an alternative label. For instance, the class ‘Book’ will have the rdfs:label ‘Material Copy of a Book’ and the alternative label ‘Book.
e.	Create a subclass of ‘Information Content Entity’ called ‘Media Content Entity’ with the definition ‘An information content entity presented in a canonical media format’.
f.	Ensure the entire IBA subclass directory is then copied under ‘Media Content Entity’ such that, for example, the class ‘Book’ is now an ICE.
g.	Add a subclass axiom on to each IBA classes which declares each ‘is carrier of’ some <corresponding media content entity>. For instance, every ‘Physical Copy of a Book’ is a carrier of some Book’.
h.	In both the IBA and the ICE hierarchy, ensure that Book, Journal Article, and Spreadsheet (in both senses) are under ‘Document’ (again, in both senses).
i.	Update the rdfs:label of ‘Document’ to ‘Document Content Entity’
j.	Rename ‘Document Form’ to ‘Form Document’
k.	Place ‘Document Field Content’ under ICE and add that is a continuant part of some Document. Add the alternative label ‘Document Field’.
l.	Move ‘Document Field’ from under IBE and place under IBA.
m.	Review all definitions of every class affected here to ensure the definitions clearly describe either ICEs or IBAs (e.g. a relationship like ‘is carrier of’ or ‘represents’ is not used inappropriately within one of these definitions).